OS X architecture auto-detection is back
This is a long overdue attempt to improve the detection of the correct
architecture flags to use when compiling the readline library on OS X.
The problem is that some of the latest compilers (e.g. MacPorts gcc 4.8.2)
only produce x86_64 code and it is therefore no longer safe to produce
i386 code by default (see issue #28).

Previous attempts looked at which -arch flags are accepted by gcc (commit
f5b6fe7) and which assemblers are installed in /usr/libexec (commit 4f74609).
Both are unreliable and were eventually removed (see issue #5).

The latest attempt compiles *and links* a Hello World with each architecture.
This could of course be done in autoconf using AC_LINK_IFELSE, but hey. This
way I'm at least allowed to set CFLAGS directly.

The use of SDK directories to detect the OS is just as bad (see issues #14,
sysroot anyway, so drop the whole business (only support 10.6+ from now on).
The LDFLAGS variable is also unnecessary.

The plan is to build the readline library with all possible architectures.
When the gnureadline extension is built using the python-config flags (which
can specify a subset of architectures), the linker will pick the appropriate
objects from the static readline library.