Remove all custom compile/link flags for extension
It is time for this bit of cargo-culting to go. The extension is already
built with the flags provided by 'python-config --cflags', which picks
the appropriate architectures and flags to match the corresponding
Python setup. There is therefore no need to override or detect this and
the module will be built as universal if needs be.

The 'strict prototypes' C warning does not have to be suppressed anymore
as the underlying cause has been fixed in recent versions of the module.

After this update, compilation no longer depends on fragile distutils
platform checks and directory name checks. This should address
issues #28, #22 and maybe even #20. Tested on 10.6 System Python and
10.9 Homebrew Python.