Use $CC or cc to compile libreadline on Darwin
Force the compiler to CC (or cc by default) to ensure that the
architecture flag detection is done on the same compiler that is used to
compile libreadline. Without this libreadline will default to gcc, which
might not be the same as cc and also does not support -arch if it is the
real gcc. The Python extension module will typically be compiled by clang
on Darwin, which should be the default cc.

This follows @ohanar's suggestion in issue #28.