Build against readline 6.3 (patch-level 3)
Use the latest and greatest readline, but only after the first
official patches have been released. Patch 002 is especially
helpful as it fixes a segfault when gnureadline is used with
IPython. This version of readline should support all future
versions of Darwin for now, so drop the 005a patch modification.