Move standard cibuildwheel variables to config
This makes it independent of the particular CI (i.e. GitHub Actions).
We don't want to get bitten again :-) We'll always e.g. need ncurses.

Also disable PyPy wheels on all platforms (it fails) and delete the
readline static library before we start each build in the hope that
it will be recompiled with universal arches when the time comes.