Restore Python 3.6 wheel support
According to https://pypistats.org/packages/numpy Python 3.6 is still
the third-most downloaded version (more than 3.9 and 3.10 combined!)...
Maybe it's because it ships with Ubuntu 18.04 LTS. Keep building wheels
for it for now.

Remove comment about the origin of this YAML since it has morphed a lot
in the meantime.