Update release procedure
The wheels are now built on GitHub Actions with cibuildwheel. Since
the Rackspace SciPy wheelhouse is no more, the old instructions won't
work anyway.

While it is certainly an option to push those wheels straight to PyPI
from GitHub Actions I prefer the manual route for now as it gives me
a chance to sign them and verify them as well.