Restrict old readline package on modern systems
The Modules/3.x/readline.c is from Python 3.3, so don't go beyond that.

Also do a basic compiler fix by dropping i386 on macOS 10.9+ and adding
arm64 on macOS 11+ (FWIW :-)). This would allow older Pythons to use
this on newer macOSes (but please don't do that!). I have verified the
combination of Python 3.3.7 and macOS 11.6.6.

Hopefully this will block readline from getting installed on modern
systems in a more sensible way, by complaining about Python instead of
crashing during compilation. I prefer this to complete removal of the
package, since someone somewhere might still depend on it.

It would be nice if pip had a way to convey a message to the user to
switch to gnureadline instead after failing on the Python version.

This fixes ludwigschwardt/python-readline#1.