Make GPLv3 license explicit
Since the Python module wraps and contains GNU readline which is GPLv3,
it makes sense to be GPLv3 too. It always has been, but this has now
been clarified by copying COPYING from readline 6.3 to LICENSE and
picking a more specific trove classifier in setup.py.

This is a duplicate of commit a96256a.