Override default readline via site customization
Over the years we have tried and debated many ways to override the
default readline module in Python so that tab completion can be fixed
in the standard interactive interpreter (see e.g. #26 and #27).

The original solution was `easy_install readline`: ugly but effective.
In the pip era this option has fallen away. The current installation
route of `pip install gnureadline` leaves it up to the user to override
`sys.modules['readline']` or to `import gnureadline as readline`.

The problem is that those tricks have to happen before the interpreter
even prints its startup banner. Python 3 roughly starts up like this:

  - run site.py:
    - set up `sys.__interactivehook__` [but DON'T import readline yet]
    - run sitecustomize.py
    - run usercustomize.py
  - import rlcompleter [imports readline!]

  - output:
  Python 3.11.3 (main, May  9 2023, 18:38:16) [...] on darwin
  Type "help", "copyright", "credits" or "license" for more information.

  - run PYTHONSTARTUP
  - run sys.__interactivehook__() [also imports readline fwiw]
  - output:
  >>>

PYTHONSTARTUP is too late. It can modify the interactive hook but not
readline itself. That leaves sitecustomize and usercustomize. It feels
a bit iffy when an installed package modifies those but, in mitigation,
the user has to push the button and the override tries to explain what
it does in some detail.

When you run `python -m override_readline`, the script first targets
sitecustomize.py because it is better for virtual environments and
system-wide overrides. If you don't have permission, it falls back to
usercustomize.py. It does `sys.modules['readline'] = gnureadline` and
also lets `sys.__interactivehook__` print a reminder that readline is
overridden (and where the override is). This is better than a direct
print() which also pops up in non-interactive Python invocations like
pip, etc.

The override is reasonably foolproof and won't crash even if gnureadline
is uninstalled and the override inadvertently remains behind. And best
of all, it all works in Python 2.7 too because site customization has
been around forever :-P

The nice thing about running a module as a script is that it inherently
uses the appropriate python interpreter, which avoids confusion. This
script could even form the core of a revamped `readline` package that
switches between, say, pyreadline and gnureadline based on the platform.