Add note to readline.py to remind of its use case
The local package version of `readline.py` is less important now that
there is an override mechanism. I kept it though because it still caters
for cases like ActiveState Python which has no readline module at all.
In those cases it is sufficient to install gnureadline and be done with
it, with no need for the override.

Add a note to explain this in case I forget again :-) And remind folks
not to go down the dark `sys.path` with this module.