Undefine the WITH_EDITLINE macro
Since Python 3.10 it is possible to build against libedit explicitly.
This introduced a WITH_EDITLINE macro in pyconfig.h which causes the
readline extension module to include the libedit headers instead of the
readline headers. Homebrew Python 3.11 finally took the plunge and
enabled this macro, causing builds of the extension code to fail with
undeclared identifiers and implicit declarations due to the missing
includes.

I first tried to add `undef_macros=['WITH_EDITLINE']` to the Extension
object in setup.py but that had no effect (probably because pyconfig.h
is included later). The next best thing is less elegant but effective:
undefine it just before it would be used.

This addresses @milosivanovic's build issues in #62.