Force local readline includes
For some reason macOS now puts /usr/include in front of the directories
listed in the `include_dirs` parameter to `Extension`. This already
contains a readline/readline.h but it is libedit in disguise.
Compilation fails with some undeclared functions like `append_history`
and `rl_resize_terminal`.

Rather force both the module and the locally checked out readline
library to include the local readline headers. This is done explicitly
for the modules and implicitly for the readline library by setting the
READLINE_LIBRARY define. This removes the need for `include_dirs` since
we are not including via angle brackets anymore (no more poor form!).