Improve reporting after successful override
Print the full path of the customisation module (it was printed when the
file had to be created from scratch but not when it was appended to).

Suggest that the file can be deleted if unwanted *only if* we created it
in the first place, to prevent possible tears :-)

Make the text print banners private to the module.