Rework unit tests to check override_readline
Drop simplistic import test. Consolidate all `import readline` checks
in a single test (`test_identity_of_readline_module`). The test module
gains two script options that will explicitly check whether readline
is or is not the same as gnureadline. Use this to check whether
`override_readline` works, by first verifying that the modules differ,
applying the override and verifying that the modules are the same
afterwards. This is all run from within the tests directory to avoid
picking up the alternative readline.py module in the project directory.