Disable Python 3.7 test on macOS 13
This combination seems to be unpopular, since the (binary) Python
distribution is not on the runner machine and has to be downloaded:

  Version 3.7 was not found in the local cache
  Version 3.7 is available for downloading
  Download from "https://github.com/actions/python-versions/releases/download/3.7.17-5356448435/python-3.7.17-darwin-x64.tar.gz"

However, the standard readline module (readline.cpython-37m-darwin.so)
cannot be imported because the ncurses library could not be found
(it's looking for /usr/local/opt/ncurses/lib/libncursesw.6.dylib
instead of the more usual /usr/lib/libncurses.5.4.dylib).

I opened an issue on GitHub Actions (actions/setup-python#859), so
this option might still come back.