Support built-in readlines in override script
The `python-build-standalone` Python distribution (used by e.g. mise)
compiles the readline module into libpython as a built-in module,
instead of the usual extension in the lib-dynload directory.

This means that the `module.__file__` attribute does not exist. Check
for this and report a fallback string instead of crashing.

This fixes #74.